Split up tests by respective extrinsics
